Added Slapping Decal Effect
Had a lot of difficulties getting it to work. There are some blocks that should work a certain way but don't have any functionality yet. This is surprising as this is an engine used by AAA studios.

Also, had to add a fade block to the texture of the decal so that it faded out. You have to actually program in the fade. This is so that developers can customize the fade in various artistic ways.

Also, when the decal was attached to a component instead of a parent actor, it lagged.

Also, a casting for the controllers was used to see which hand slapped and scaling was used to flip the decal. This is currently branched into right and left. This could late be streamlined, but seems to work fine.

Sound was added as well, but will need to test later if sound is played at full level for everyone or if it's just local to the vicinity of the slap point

Helpful Links
https://answers.unrealengine.com/questions/459978/bug-decal-fade-out-doesnt-work.html