Added Slapping Effect
Added a visible effect that appears when slapping. It spawns on the collision sphere. Modified built in Niagara particle effect. Note that the ribbon is in the 2nd block on the editor. This caused confusion when trying to reduce the ribbon radius.

Also, added SlapSprayContainer so that the person getting hit can't see the effect, because it doesn't look good.